+++ beetsplug/lyrics.py
@@ -329,10 +329,10 @@ def _scrape_strip_cruft(html, plain_text_out=False):
     """
     html = unescape(html)
 
-    # Normalize EOL
-    html = html.replace('\r', '\n')
+    html = html.replace('\r', '\n')  # Normalize EOL.
     html = re.sub(r' +', ' ', html)  # Whitespaces collapse.
-    html = BREAK_RE.sub('\n', html)  # <br> eats up surrounding '\n'
+    html = BREAK_RE.sub('\n', html)  # <br> eats up surrounding '\n'.
+    html = re.sub(r'<(script).*?</\1>(?s)', '', html)  # Strip script tags.
 
     if plain_text_out:  # Strip remaining HTML tags
         html = TAG_RE.sub('', html)